+++ CHANGES.md
@@ -1,8 +1,18 @@
 # Changelog
 
+## 0.3.0
+
+### Added
+
+- Added NOVOplasty mediated mitochondrial assembly, `AAFTF mito`
+- integrated `AAFTF mito` into `AAFTF pipeline`
+- updated sourmash LCA database link and will download if not found
+
+
+
 ## 0.2.5
 
-### Added 
+### Added
 
 - Issue #10 added GC% in the assessment report table
 - Added --mem option to pilon to up the heapsize for java runs